feat(featured-item): rename featured-tile to featured-item
