+++ components/index.js
@@ -32,7 +32,7 @@ export { default as ConfirmationDialogTemplate } from './confirmation-dialog-tem
 export { default as ConfirmDrawer } from './confirm-drawer';
 export { default as ContentWrapper } from './content-wrapper';
 export { default as Drawer } from './drawer';
-export { default as FeaturedTile } from './featured-tile.vue';
+export { default as FeaturedItem } from './featured-item.vue';
 export { default as FeatureHighlight } from './feature-highlight';
 export { default as FeatureHighlightRow } from './feature-highlight-row';
 export { default as GuestInformationForm } from './guest-information-form';
