+++ stories/components/basic/featured-item.js
@@ -1,22 +1,22 @@
-import FeaturedTile from '../../../components/featured-tile.vue';
+import FeaturedItem from '../../../components/featured-item.vue';
 
 export default {
-	title: 'Components / Basic / FeaturedTile',
-	component: FeaturedTile,
+	title: 'Components / Basic / FeaturedItem',
+	component: FeaturedItem,
 };
 
 export function Overview() {
 	return {
-		components: { FeaturedTile },
-		template: '<featured-tile :tile="tile" @selected="handleSelected" />',
+		components: { FeaturedItem },
+		template: '<featured-item :item="item" @selected="handleSelected" />',
 		data() {
 			return {
-				tile: {
+				item: {
 					id: '1234567890',
 					type: 'category',
 					imageUrl: 'https://d3e4ixgvd0ibii.cloudfront.net/shopify/pages/home/home__cleanse-tile.jpg',
-					title: 'Test Tile',
-					description: 'This is a description of the test tile.',
+					title: 'Item Title',
+					description: 'Item description',
 				},
 			};
 		},
